Aspen DB remove request metadata
The import file for the Aspen Database previously included all the
metadata for the relay and request, which created a bug in which the
request date and status would be over-written on import. With this
update, only the minimum metadata needed to allow the settings to import
is included in the import file.