Not include strikeout, exclude blank cells
The logic to exclude text in strikeout format was expanded to everywhere
text is filtered on style, and the logic was pulled out to a function.
Also, cells in multi-line setting tables (e.g. protection & automation
logic) where the cell has no setting is skipped rather than counted as a
blank setting line.